+++ src/components/Dashboard/TeamViewer/Explore.jsx
@@ -1,0 +1,101 @@
+import React, { useState, useEffect } from "react";
+import { List, Grid, Typography } from "@material-ui/core";
+import Pagination from "@material-ui/lab/Pagination";
+
+import PropTypes from "prop-types";
+import TeamLoading from "../TeamLoading";
+
+import RenderRow from "./RenderRow";
+
+function Explore(props) {
+    const [matches, setMatches] = useState({});
+    const [originalTeamId, setOriginalTeam] = useState({});
+    const [page, setPage] = React.useState(1);
+    const [allTeams, setAllTeams] = useState({});
+    const [loading, setLoading] = useState("Loading your matches...");
+    useEffect(() => {
+        props.profile.getTeamUser().then((success) => {
+            const team_id = success.response.team_id;
+            setOriginalTeam(success.response.team_id);
+            props.profile.matches(team_id).then((success) => {
+                setMatches(success.response);
+                setLoading(false);
+            });
+        });
+        props.profile.getAllTeams(0, 4).then((success) => {
+            setLoading(false);
+            setAllTeams(success.response);
+        });
+    }, []);
+    const handlePagination = (event, value) => {
+        setPage(value);
+        props.profile.getAllTeams(((value - 1) * 4), 4).then((success) => {
+            setLoading(false);
+            setAllTeams(success.response);
+        });
+    };
+    if (loading) {
+        return (<TeamLoading text={loading} />);
+    }
+    return (
+        <Grid item
+            container
+            direction="column"
+            justify="center"
+            alignItems="center">
+            <Typography variant="h5"
+                style={{ marginTop: "2em" }}>
+                Matches
+            </Typography>
+            <List
+                style={{ maxHeight: "300px", width: "600px", overflow: "auto" }}
+                className="no-scrollbars no-style-type"
+            >
+                {matches.matches && matches.matches.length > 0 ? (
+                    matches.matches.map((invitingTeamId, i) => (
+                        <RenderRow
+                            key={i}
+                            invitingTeam={invitingTeamId}
+                            originalTeamId={originalTeamId}
+                            {...props}
+                        />
+                    ))
+                ) : (
+                    <Typography variant="subtitle1">No Matches Yet</Typography>
+                )}
+            </List>
+            <Typography variant="h5">Explore</Typography>
+            <List
+                style={{ maxHeight: "300px", width: "600px", overflow: "auto" }}
+                className="no-scrollbars no-style-type"
+            >
+                {allTeams.all_open_teams && allTeams.all_open_teams.length > 0 ? (
+                    allTeams.all_open_teams.map((invitingTeamId, i) => (
+                        <RenderRow
+                            key={i}
+                            invitingTeam={invitingTeamId}
+                            originalTeamId={originalTeamId}
+                            {...props}
+                        />
+                    ))
+                ) : (
+                    <Typography variant="subtitle1">No Teams Yet</Typography>
+                )}
+            </List>
+            <Pagination
+                count={Math.ceil(allTeams.total_teams / 4)}
+                variant="outlined"
+                page={page}
+                onChange={handlePagination}
+                shape="rounded"
+            />
+        </Grid>
+    );
+}
+Explore.propTypes = {
+    invitingTeam: PropTypes.object,
+    originalTeamId: PropTypes.string,
+    profile: PropTypes.object,
+};
+
+export default Explore;
