+++ src/components/Dashboard/TeamViewer/RenderRow.jsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { ListItemSecondaryAction, IconButton, ListItemAvatar, ListItem, ListItemText, Avatar } from "@material-ui/core";
+import PropTypes from "prop-types";
+import GroupAdd from "@material-ui/icons/GroupAdd";
+
+function RenderRow(props) {
+    const { invitingTeam, originalTeamId } = props;
+    return (
+        <ListItem style={{ padding: "1em", cursor: "pointer" }}>
+            <ListItemAvatar>
+                <Avatar>{invitingTeam.name ? invitingTeam.name.substring(0, 1) : ""}</Avatar>
+            </ListItemAvatar>
+            <ListItemText
+                primary={invitingTeam.name ? invitingTeam.name : ""}
+                secondary={`${invitingTeam.members ? invitingTeam.members.length : ""}/4`}
+            />
+            <ListItemSecondaryAction>
+                <IconButton
+                    edge="end"
+                    aria-label="add"
+                    onClick={() => props.profile.inviteTeam(originalTeamId, invitingTeam.team_id)}
+                >
+                    <GroupAdd />
+                </IconButton>
+            </ListItemSecondaryAction>
+        </ListItem>
+    );
+}
+RenderRow.propTypes = {
+    invitingTeam: PropTypes.object,
+    originalTeamId: PropTypes.string,
+    profile: PropTypes.object
+};
+
+export default RenderRow;
